reafactor : 피드백 반영 (Repolmpl 생성 후 bind)
