+++ app/src/main/res/values/strings.xml
@@ -4,7 +4,7 @@
     <string name="searchButton">검색</string>
     <string name="noSearch">검색 결과가 없습니다</string>
     <string name="x">X</string>
-    <string name="kakao_api_key">23f7816cce557c3510b52b6418462b31</string>
+    <string name="kakao_api_key">9fb498a1684eadf092d3f0e74ffac826</string>
     <string name="errorMessage">지도 인증에 실패했습니다.\n다시 시도해 주세요.</string>
     <string name="retry">재시도</string>
 </resources>
